Refactor build system and update documentation
- Removed outdated files including `BUILD_README.md`, `build-optimized.js`, and various scripts to streamline the project structure.
- Introduced `build-classic.js` for a classic build approach, maintaining compatibility with previous configurations.
- Updated `.gitignore` to include `node_modules` and ensure cleaner project management.
- Enhanced `README.md` and `QUICKSTART.md` for clearer setup instructions and performance optimizations.
- Updated `ObjectBuilder-app.xml` for improved performance settings and copyright information.
