+++ QUICKSTART.md
@@ -1,0 +1,111 @@
+# 🚀 ObjectBuilder - Quick Start Guide
+
+**Get up and running with the optimized ObjectBuilder in under 5 minutes!**
+
+## ⚡ Instant Launch (Recommended)
+
+**Just run the optimized executable:**
+
+```bash
+bin\ObjectBuilder.exe\ObjectBuilder.exe
+```
+
+**For maximum performance:**
+
+```bash
+.\launch-optimized.bat
+```
+
+That's it! The optimized version is ready to use.
+
+## 🔧 Development Setup (Optional)
+
+### Prerequisites
+
+- **Node.js 16+** ([Download](https://nodejs.org/))
+- **Windows 10/11**
+
+### Setup Commands
+
+```bash
+# Install dependencies
+npm install
+
+# Setup SDKs (first time only)
+node setup.js
+
+# Build optimized version
+npm run build:classic
+```
+
+## 📦 Available Commands
+
+| Command                  | Description              | Use Case             |
+| ------------------------ | ------------------------ | -------------------- |
+| `npm run build`          | Build optimized version  | Production build     |
+| `.\package-simple.bat`   | Package executable       | Create .exe bundle   |
+| `.\launch-optimized.bat` | Maximum performance mode | Ultimate performance |
+| `npm run run`            | Launch executable        | Test final version   |
+| `npm run debug`          | Debug mode               | Development/testing  |
+| `npm run clean`          | Clean build files        | Fresh start          |
+
+## 🎯 Performance Scripts
+
+### **🚀 Maximum Performance**
+
+```bash
+.\launch-optimized.bat
+```
+
+- Forces high performance power profile
+- DirectX hardware acceleration
+- Optimized runtime flags
+
+### **📦 Simple Packaging**
+
+```bash
+.\package-simple.bat
+```
+
+- Automated executable creation
+- Correct icon integration
+- Error handling
+
+## 🎯 Performance Features
+
+✅ **DirectX Hardware Acceleration**  
+✅ **60+ FPS Performance**  
+✅ **Smooth Scrolling**  
+✅ **Reduced Memory Usage**  
+✅ **Faster Startup**
+
+## 🛠️ File Support
+
+- **.dat** files - Tibia metadata
+- **.spr** files - Tibia sprites
+- **.obd** files - ObjectBuilder projects
+
+## 📋 Quick Troubleshooting
+
+**Q: Won't start?**
+
+- Update DirectX drivers
+- Try: `.\debug-run.bat`
+
+**Q: Poor performance?**
+
+- Use: `.\launch-optimized.bat`
+- Close other applications
+
+**Q: Build fails?**
+
+- Run: `node setup.js`
+- Check disk space (>1GB)
+
+## 🎉 You're Ready!
+
+The optimized ObjectBuilder delivers **superior performance** compared to the original. Enjoy the smooth, responsive experience!
+
+---
+
+**🚀 Pro Tip**: Use `.\launch-optimized.bat` for the absolute best performance!
